Refactored Project Structure after branch merging
