+++ app/src/main/java/com/example/unilink/Fragments/Registration/addProfileBannerFragment.java
@@ -1,17 +1,17 @@
-package com.example.unilink;
+package com.example.unilink.Fragments.Registration;
 
 import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentTransaction;
 
 import android.provider.MediaStore;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.ImageButton;
+
+import com.example.unilink.R;
 
 /**
  * A simple {@link Fragment} subclass.
