+++ app/src/main/java/com/example/unilink/Models/UnilinkUser.java
@@ -1,179 +1,84 @@
 package com.example.unilink.Models;
 
-import android.content.Context;
-import android.os.Parcel;
-import android.os.Parcelable;
+import com.example.unilink.Models.Interests.Food;
+import com.example.unilink.Models.Interests.Gaming;
+import com.example.unilink.Models.Interests.InterestsCategory;
+import com.example.unilink.Models.Interests.Movie;
+import com.example.unilink.Models.Interests.Music;
+import com.example.unilink.Models.Interests.Sport;
 
-import com.google.firebase.Timestamp;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.PriorityQueue;
+import java.time.LocalDateTime;
 
-import java.io.Serializable;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Random;
-import java.util.UUID;
+public class UnilinkUser {
 
-public class UnilinkUser implements Parcelable, Serializable {
-    private String Uid;
-    private String auth_uid;
-    private HashMap<String, String> user_fullName; // first name and last name
-    private String user_phoneNum;
-    private String user_email;
-    private Timestamp user_lastUpdated;
+    private String userID;
+    private String bio;
+    private LocalDateTime timeCreated;
+    public List<String> ConnectedUIDs;
+    private PriorityQueue<InterestsCategory> categories;
 
-    // private String user_bio; // not applied just yet
-    // private String user_profilepic; // not applied just yet
-
-    public UnilinkUser() {
-        this.user_email = null;
-        this.user_fullName = new HashMap<String, String>();
-        this.user_phoneNum = null;
+    public UnilinkUser(String userID, String bio) {
+        this.userID = userID;
+        this.bio = bio;
+        this.categories = new PriorityQueue<>();
+        categories.add(new Food(0));
+        categories.add(new Music(0));
+        categories.add(new Movie(0));
+        categories.add(new Sport(0));
+        categories.add(new Gaming(0));
+        this.ConnectedUIDs = new ArrayList<>();
+        this.timeCreated = LocalDateTime.now();
     }
 
-    public UnilinkUser(String auth_id, String firstName, String lastName, String user_phoneNum, String user_email, String uid) {
-        // add user Id from Firebase
-        this.auth_uid = auth_id;
-        // creating first name and full name
-        this.user_fullName = new HashMap<String, String>();
-        user_fullName.put("firstName", firstName.substring(0,1).toUpperCase() + firstName.substring(1));
-        user_fullName.put("lastName", lastName.substring(0,1).toUpperCase() + lastName.substring(1));
-
-        // adding in phonenumber and email
-        this.user_phoneNum = user_phoneNum;
-        this.user_email = user_email;
-        this.user_lastUpdated = Timestamp.now();
-        this.Uid = (uid != null ? uid : UUID.randomUUID().toString());
+    public PriorityQueue<InterestsCategory> getCategories() {
+        return categories;
     }
 
-    public UnilinkUser getDataFrom(UnilinkUser usr) {
-        this.auth_uid = usr.auth_uid;
-        this.user_fullName = usr.user_fullName;
-        this.user_email = usr.user_email;
-        this.user_phoneNum = usr.user_phoneNum;
-        this.user_lastUpdated = usr.user_lastUpdated;
-        this.Uid = usr.getUid();
-        return this;
+    public String getBio() {
+        return bio;
     }
 
-    @Override
-    public String toString() {
-        return "[UID: " + Uid + "; Email: " + this.user_email + "; FullName: " + this.user_fullName + " ; Phone: " + this.user_phoneNum + "]";
+    public void setBio(String bio) {
+        this.bio = bio;
     }
 
-    /* #region getName */
-    // get the full name in one string
-    public String getFullName() {
-        if (this.user_fullName.isEmpty())
-            return null;
-
-        String fN = this.user_fullName.get("firstName");
-        String lN = this.user_fullName.get("lastName");
-
-        if (fN != null && lN != null)
-            return fN + " " + lN;
-        else
-            return null;
-    }
-
-    // Getting firstName information
-    public String getFirstName() {
-        if (this.user_fullName.isEmpty())
-            return null;
-        String fN = this.user_fullName.get("firstName");
-        if (fN != null)
-            return fN.substring(0,1).toUpperCase() + fN.substring(1);
-        else
-            return null;
-    }
-
-    public String getLastName() {
-        if (this.user_fullName.isEmpty())
-            return null;
-        String lN = this.user_fullName.get("lastName");    
-        if (lN != null)
-            return lN.substring(0,1).toUpperCase() + lN.substring(1);
-        else
-            return null;
-    }
-    /* #endregion */
-
-    /* #region getOtherInfo */
-    public String getPhoneNum() {
-        return this.user_phoneNum;
-    }
-
-    public String getEmail() {
-        return this.user_email;
-    }
-
-    public String getAuthId() {
-        return this.auth_uid;
-    }
-
-    public String getUid() {return this.Uid;}
-    /* #endregion */
-
-    /* #region setters */
-    public void setFirstName(String fN) {
-        this.user_fullName.put("firstName", fN);
-    }
-
-    public void setLastName(String lN) {
-        this.user_fullName.put("lastName", lN);
-    }
-
-    public void setEmail(String email) {
-        this.user_email = email;
-    }
-
-    public void setPhoneNum(String pN) {
-        this.user_phoneNum = pN;
-    }
-
-    public void setAuthId(String uid) {
-        this.auth_uid = uid;
-    }
-
-    public void setFullName(String ffN) {
-        return;
-    }
-
-    public void setUid(String uid) {this.Uid = uid;}
-    /* #endregion */
-
-    /* #region Parcel Code */
-    @Override
-    public int describeContents() {
-        return 0;
-    }
-
-    // The implementation created by the Parcelabler site.
-    @Override
-    public void writeToParcel(Parcel dest, int flags) {
-        dest.writeString(this.auth_uid);
-        dest.writeString(this.user_phoneNum);
-        dest.writeString(this.user_email);
-        dest.writeSerializable(this.user_fullName);
-        dest.writeString(this.Uid);
-    }
-
-    private UnilinkUser(Parcel in) {
-        this.auth_uid = in.readString();
-        this.user_phoneNum = in.readString();
-        this.user_email = in.readString();
-        this.user_fullName = (HashMap<String, String>) in.readSerializable();
-        this.Uid = in.readString();
-    }
-
-    public static final Parcelable.Creator<UnilinkUser> CREATOR = new Parcelable.Creator<UnilinkUser>() {
-        @Override
-        public UnilinkUser createFromParcel(Parcel source) {
-            return new UnilinkUser(source);
+    public void addChosenInterest(Enum interest, InterestsCategory category) {
+        if (category instanceof Food) {
+            ((Food) category).addFoodInterest((Food.Interest) interest);
+            categories.remove(category);
+            category.setPriorityLevel(1);
+            categories.add(category);
+        } else if (category instanceof Music) {
+            ((Music) category).addMusicInterest((Music.Interest) interest);
+            categories.remove(category);
+            category.setPriorityLevel(1);
+            categories.add(category);
+        } else if (category instanceof Movie) {
+            ((Movie) category).addMovieInterest((Movie.Interest) interest);
+            categories.remove(category);
+            category.setPriorityLevel(1);
+            categories.add(category);
+        } else if (category instanceof Sport) {
+            ((Sport) category).addSportInterest((Sport.Interest) interest);
+            categories.remove(category);
+            category.setPriorityLevel(1);
+            categories.add(category);
+        } else if (category instanceof Gaming) {
+            ((Gaming) category).addGamingInterest((Gaming.Interest) interest);
+            categories.remove(category);
+            category.setPriorityLevel(1);
+            categories.add(category);
         }
+    }
 
-        @Override
-        public UnilinkUser[] newArray(int size) {
-            return new UnilinkUser[size];
-        }
-    };
-    /* #endregion */
+    public void addConnectedUser(String userID) {
+        ConnectedUIDs.add(userID);
+    }
+
+    public List<String> getConnectedUIDs() {
+        return ConnectedUIDs;
+    }
 }
