+++ app/src/main/res/layout/fragment_add_bio.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".addBioFragment">
+    tools:context=".Fragments.Registration.addBioFragment">
 
     <!-- TODO: Update blank fragment layout -->
