+++ app/src/main/res/layout/fragment_add_birthday.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".addBirthdayFragment">
+    tools:context=".Fragments.Registration.addBirthdayFragment">
 
     <!-- TODO: Update blank fragment layout -->
