+++ app/src/main/res/layout/fragment_add_profile_interest.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".addProfileInterest">
+    tools:context=".Fragments.Registration.addProfileInterest">
 
     <!-- TODO: Update blank fragment layout -->
