+++ app/src/main/res/layout/fragment_add_profile_picture.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".addProfilePictureFragment">
+    tools:context=".Fragments.Registration.addProfilePictureFragment">
 
     <!-- TODO: Update blank fragment layout -->
 
@@ -43,7 +43,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             app:srcCompat="@drawable/profileadd"
-            android:background="null"
+            android:background="@null"
             android:layout_below="@id/addyourphotodetailstextview"
             android:layout_centerHorizontal="true"
             android:layout_marginTop="64dp"
